Fixed Windows compiler warnings
git-svn-id: https://valelab.ucsf.edu/svn/micromanager2/trunk@6157 d0ab736e-dc22-4aeb-8dc9-08def0aa14fd

+++ DeviceAdapters/DemoCamera/DemoCamera.cpp
@@ -1942,13 +1942,13 @@ int CDemoObjectiveTurret::OnTrigger(MM::PropertyBase* pProp, MM::ActionType eAct
       std::string tr;
       pProp->Get(tr);
       if (tr == " + ") {
-         if (sequenceIndex_ < sequence_.size()) {
+         if (sequenceIndex_ < (int) sequence_.size()) {
             std::string state = sequence_[sequenceIndex_];
             int ret = SetProperty("State", state.c_str());
             if (ret != DEVICE_OK)
                return ERR_IN_SEQUENCE;
             sequenceIndex_++;
-            if (sequenceIndex_ >= sequence_.size()) {
+            if (sequenceIndex_ >= (int) sequence_.size()) {
                sequenceIndex_ = 0;
             }
          } else
